CLI Installer BaseCode Before AnyMods...
This is working Menu System which drives installs, and seeding from standard SQL queries which have to be written first for your SQL engine. It only uses EF core. It requires a database context. The nice thing in a hosted system the database context can be a remote IP system.