Build 15.6 Security Changes, Formatting
Updated the main App to include Usersession Logs which are required to keep the platform safe, 2 factor authentication and all future security upgrades. An Error in the SQLITE DB on the name of a field was the cause of old code not working. Suspected malicious action. UI Icon changes, and other formatting and improved mobile UI rendering are the upsides. Expecting Last Release as everything else contemplated is back office changes for now.